==> globalprep/spp/v2021/_setup/5b_generate_spp_map_list.Rmd <==
---
title: 'Generate IUCN spp map list'
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    code_folding: hide
    toc: true
    toc_depth: 3
    toc_float: yes
    number_sections: true
    theme: cerulean
    highlight: haddock
    includes: 
      in_header: '../../../../workflow/templates/ohi_hdr.html'
  pdf_document:
    toc: true
---



```{r}
knitr::opts_chunk$set(echo = TRUE, eval=FALSE)
```

``` {r setup, echo = TRUE, message = FALSE, warning = FALSE}

knitr::opts_chunk$set(fig.width = 6, fig.height = 4, fig.path = 'Figs/',
                      echo = TRUE, message = FALSE, warning = FALSE)

library(raster)
library(sf)
library(fasterize)
library(dplyr)
library(rgdal)
library(here)
library(ggplot2)
library(rgeos)

source(here('workflow/R/common.R'))

goal     <- 'spp'
scenario <- 'v2022'
dir_goal <- here('globalprep', goal, scenario)

### goal specific folders and info
dir_data  <- file.path(dir_goal, '_data')
dir_setup <- file.path(dir_goal, '_setup')
dir_anx  <- file.path(dir_M, 'git-annex/globalprep')
dir_goal_anx <- file.path(dir_anx, goal, scenario, 'spp_risk_dists')
dir_raw_data_anx <- file.path(dir_anx, '_raw_data')

source(file.path(dir_setup, 'common_fxns.R'))

dir_bli <- file.path(dir_M, 'git-annex/globalprep/_raw_data/birdlife_intl/d2021')
dir_shp <- file.path(dir_M, 'git-annex/globalprep/_raw_data/iucn_spp/d2022-1')
  ### in this folder are shapefiles at a taxonomic level, both from
  ### IUCN spatial data download (version 2022-1)

```

# Summary

* Subpopulation polygons must be identified and rasterized separately from the parent polygon; this must be done by sciname and subpop fields since the polygon IDs are based upon the parent ID.
* Regional assessments need not be determined at this stage - the ID numbers match the global ID numbers (including subpops).

# Data source

IUCN Red List: Spatial Data Download

BirdLife International 

# Methods

## Generate species map list with subpops

### Create subpopulation lookup to match maps with IUCN SIDs for subpops

From available shapefiles, reclassify subpopulation polygons into subpop ID numbers for matching to risk and trend. 

``` {r convert BLI to shp-dbf-etc}

### convert BirdLife Int'l .gdb to shapefile, after filtering to species in the "marine" list.
### Here we also standardize the column names to match the IUCN column names.

bli_marine_shp <- file.path(dir_bli, 'bli_marine_v2020.shp')

if(!file.exists(bli_marine_shp)) {
  
  bli_sf <- st_read(file.path(dir_bli, 'BOTW.gdb'))
  
  marine_spp <- read_csv(file.path(dir_data, sprintf('spp_marine_from_api_%s.csv', api_version)),
                         col_types = 'dcc')
  
  bli_marine_sf <- bli_sf %>%
  filter(SISID %in% marine_spp$iucn_sid) %>%
    clean_df_names()
  
  #if the above code ^^ does not work, use this. It functions the same way, just more tedious. 
  #marine_colnames <- colnames(marine_spp[2:3])
  #bli_marine_sf_check <- bli_sf %>%
  #  inner_join(marine_spp, by = c("SISID" = "iucn_sid")) %>%
  #  dplyr::select(-marine_colnames) %>%
  #    clean_df_names() 
  
  st_geometry(bli_marine_sf) <- NULL #gets rid of "Shape" column. Since geometry is sticky, clean_df_names() retains the original geometry column "Shape". Now lets rename it to "shape" and retain the geometry format.
  st_geometry(bli_marine_sf) <- bli_marine_sf$shape  #have to reestablish that "shape" is the geometry column. For some reason renaming the geometry column results in r not recognizing the column as the geometry column. 
  
  st_write(bli_marine_sf, bli_marine_shp)
}

```

Note that there appear to be no subpopulation polygons within the BirdLife International map set; there is no "subpop" column in that attribute table.  It appears that the individual files may not have "subpop" column either.

``` {r create_subpop_lookup}
iucn_dbfs <- list.files(dir_shp, pattern = '\\.dbf$', full.names = TRUE)
bli_dbfs  <- list.files(dir_bli, pattern = '\\.dbf$', full.names = TRUE)
shps_dbfs <- c(iucn_dbfs, bli_dbfs)

### This is all "marine" species by habitat and system:
marine_spp <- read_csv(file.path(dir_data, 
                                 sprintf('spp_marine_from_api_%s.csv', 
                                         api_version)),
                       col_types = 'dcc')

### This is *all* spp from the API, filtered down to those in marine_spp: 
api_spp <- read_csv(file.path(dir_goal_anx, 'iucn', 
                              sprintf('spp_info_from_api_%s.csv', api_version)),
                    col_types = 'dccccccccc') %>%
  filter(iucn_sid %in% marine_spp$iucn_sid)

### Define helper function to cleanly get the info from DBF files,
### including colname checks
get_dbf <- function(x) { ###x <- shps_dbfs[1]
  spp_dbf_info <- foreign::read.dbf(x, as.is = TRUE) %>%
    clean_df_names()
  
  ### Individual files call the id number 'iucn_sid'; bli has been fixed to
  ### 'iucn_sid' as well; Red List Data Download files are 'id_no'... So:
  ### fix that here!
  names(spp_dbf_info)[names(spp_dbf_info) == 'id_no'] <- 'iucn_sid'

  ### if 'binomial', set to 'sciname'
  if(!'sciname' %in% names(spp_dbf_info)) {
    names(spp_dbf_info)[names(spp_dbf_info) == 'binomial'] <- 'sciname'
  }
  
  ### Some files (e.g. bli and individual files) don't have 'subpop':
  ### add it in as NAs.
  if(!'subpop' %in% names(spp_dbf_info)) spp_dbf_info$subpop <- NA
  
  ### some files don't have 'presence' field: if not present or NA,
  ### set to presence = 1.  Some have presence = 0?
  if(!'presence' %in% names(spp_dbf_info)) spp_dbf_info$presence <- 1
  
  spp_dbf_info <- spp_dbf_info %>%
    mutate(presence = ifelse(presence == 0, 1, presence))
  
  return(spp_dbf_info)
}

## only run on IUCN here... 
shps_df <- lapply(iucn_dbfs, get_dbf) %>%
  setNames(iucn_dbfs) %>%
  bind_rows(.id = 'dbf_file') %>%
  dplyr::select(dbf_file, iucn_sid, sciname, presence, subpop) %>%
  filter(iucn_sid %in% marine_spp$iucn_sid)

## now run on birdlife data here, and change column names to match
 x <- bli_dbfs
  spp_dbf_info <- foreign::read.dbf(x, as.is = TRUE) %>%
    clean_df_names()
  
  colnames(spp_dbf_info)
#    [1] "sisid"   "binomil" "presenc" "origin"  "seasonl" "compilr" "yrcmpld" "citatin" "source"  "dst_cmm" "version"
# [12] "shp_lng" "shape_r"
  
  ### Individual files call the id number 'iucn_sid'; bli is named sisid
  ### 'iucn_sid' as well;
  ### fix that here!
  names(spp_dbf_info)[names(spp_dbf_info) == 'sisid'] <- 'iucn_sid'

  ### if 'binomial', set to 'sciname'
  if(!'sciname' %in% names(spp_dbf_info)) {
    names(spp_dbf_info)[names(spp_dbf_info) == 'binomil'] <- 'sciname'
  }
  
  ### Some files (e.g. bli and individual files) don't have 'subpop':
  ### add it in as NAs.
  if(!'subpop' %in% names(spp_dbf_info)) spp_dbf_info$subpop <- NA
  
  ## fix presenc colname in bli to be presence
  names(spp_dbf_info)[names(spp_dbf_info) == 'presenc'] <- 'presence'
  
  ### some files don't have 'presence' field: if not present or NA,
  ### set to presence = 1.  Some have presence = 0?
  if(!'presence' %in% names(spp_dbf_info)) spp_dbf_info$presence <- 1
  
  spp_dbf_info <- spp_dbf_info %>%
    mutate(presence = ifelse(presence == 0, 1, presence))


shps_bli_df <- spp_dbf_info %>%
  cbind(bli_dbfs) %>%
  dplyr::select("dbf_file" = "bli_dbfs", iucn_sid, sciname, presence, subpop) %>%
  filter(iucn_sid %in% marine_spp$iucn_sid)  
  
## now rbind 

shps_df <- rbind(shps_df, shps_bli_df)

shp_subpops <- shps_df %>%
  dplyr::select(shp_iucn_sid = iucn_sid, sciname, shp_subpop = subpop) %>%
  filter(!is.na(shp_subpop)) %>%
  distinct() %>%
  mutate(shp_subpop_clean = str_replace(tolower(shp_subpop), ' subpopulation| ocean', ''),
         shp_subpop_clean = str_replace_all(shp_subpop_clean, '[^a-z]', ' '),
         shp_subpop_clean = str_trim(shp_subpop_clean),
         shp_subpop_clean = str_replace(shp_subpop_clean, 'noth', 'north'))
           ### fix typo

api_subpops <- api_spp %>%
  dplyr::select(api_iucn_sid = iucn_sid, sciname, api_subpop = population) %>%
  filter(!is.na(api_subpop)) %>%
  filter(sciname %in% shp_subpops$sciname) %>%
  distinct() %>%
  mutate(api_subpop_clean = str_replace_all(tolower(api_subpop), ' subpopulation| ocean', ''),
         api_subpop_clean = str_replace_all(api_subpop_clean, '[^a-z]', ' ') %>% str_trim)

subpops_match_raw <- shp_subpops %>%
  full_join(api_subpops, by = 'sciname') %>%
  group_by(api_iucn_sid) %>%
  mutate(subpop_match = str_detect(shp_subpop_clean, api_subpop_clean),
         n_match = sum(subpop_match)) %>%
  filter(subpop_match | sum(subpop_match) == 0) %>%
  ungroup()

caretta_subpops <- subpops_match_raw %>%
  filter(shp_iucn_sid == 3897) %>%
  mutate(api_subpop_single = str_split(api_subpop_clean, ' ')) %>%
  unnest(api_subpop_single) %>%
  group_by(api_subpop_clean, shp_subpop_clean) %>%
  mutate(match = str_detect(shp_subpop_clean, api_subpop_single),
         n_match = sum(match),
         n_words = n()) %>%
  filter(sum(match) == n()) %>%
  ungroup() %>%
  dplyr::select(shp_iucn_sid, sciname, shp_subpop, api_iucn_sid, api_subpop) %>%
  distinct()

subpops_match <- subpops_match_raw %>%
  filter(shp_iucn_sid != 3897 & subpop_match == TRUE) %>%
  dplyr::select(shp_iucn_sid, sciname, shp_subpop, api_iucn_sid, api_subpop) %>%
  bind_rows(caretta_subpops)

### Check for missed matches:

# api_subpops$api_iucn_sid[!api_subpops$api_iucn_sid %in% subpops_match$api_iucn_sid]
### API subpop, not matched: 16369383 Tursiops truncatus 
###                                   Mediterranean subpopulation
### * OK - no polygon for this subpopulation.  Regional assessment?
### API subpop, not matched: 44187 Orcaella brevirostris 
###                                Malampaya Sound subpopulation
### * Doesn't look like there is a shape for this subpop... Sorry dude!
### API subpop, not matched: 123095978 Orcaella brevirostris
###                                    loilo-Guimaras Subpopulation  
### * Doesn't look like there is a shape for this subpop... Sorry dude!
### API subpop, not matched: 156206333  Delphinus delphis
###                                    Gulf of Corinth subpopulation
### * It does look like there is a shape for this one... the medittarean sea.. but it seems like it is already accounted for in the matched subpops... so i think it is ok. 

 shp_subpops[!shp_subpops$shp_subpop %in% subpops_match$shp_subpop, ]
# all matched
 
write_csv(subpops_match, 
          file.path(dir_setup, sprintf('int/subpops_match_api_to_shp_%s.csv', 
                                     api_version)))

```

### get info on comprehensive assessments

Not all maps are for species that are included in comprehensively assessed taxonomic groups.  Get info on comprehensive assessment status for species from the API.

``` {r get comp assess status}
comp_file <- file.path(dir_data, 
                       sprintf('iucn_comp_assessed_%s.csv', api_version))
if(!file.exists(comp_file)) {
  comp_gps_url <- 'http://apiv3.iucnredlist.org/api/v3/comp-group/list?token=%s'
  comp_gps <- fromJSON(sprintf(comp_gps_url, api_key)) %>%
    .$result
  
  comp_spp_url <- 'http://apiv3.iucnredlist.org/api/v3/comp-group/getspecies/%s?token=%s'
  comp_spp <- mc_get_from_api(comp_spp_url, comp_gps$group_name, api_key) %>%
    dplyr::select(iucn_sid = taxonid, sciname = scientific_name)
  
  write_csv(comp_spp, comp_file)
} else {
  comp_spp <- read_csv(comp_file,
                       col_types = 'dc')
}
```

### Join map list to subpop corrected list; remove duplicated spp IDs

From the species range map shapefiles, pull map info from the .dbf files.  For species listed in multiple files (e.g. sea snakes and reptiles), remove duplicates.  For species with polygons differentiated by subpopulation, adjust the iucn_sid to match the subpop info.

Add a flag for comprehensively-assessed spp - fix BONEFISH_TARPONS to all be assessed - either it's comprehensive or not - and add sea turtle species (6 of 7 assessed).

``` {r generate map list from dbfs}

iucn_dbfs  <- list.files(dir_shp, pattern = '\\.dbf$', full.names = TRUE)
bli_dbfs   <- list.files(dir_bli, pattern = '\\.dbf$', full.names = TRUE)

map_files  <- c(iucn_dbfs, bli_dbfs)


spp_group_names <- map_files %>%
  basename() %>%
  str_replace_all('\\.dbf$|_PART_.+', '') %>%
  unique()


####### just map iucn files
map_info_raw_iucn <- lapply(iucn_dbfs, get_dbf) %>%
  setNames(iucn_dbfs) %>%
  bind_rows(.id = 'dbf_file') %>%
  dplyr::select(shp_iucn_sid = iucn_sid, sciname, presence, subpop, dbf_file)

####### now map just the bli files
## now run on birdlife data here, and change column names to match
 x <- bli_dbfs
  spp_dbf_info <- foreign::read.dbf(x, as.is = TRUE) %>%
    clean_df_names()
  
  colnames(spp_dbf_info)
#    [1] "sisid"   "binomil" "presenc" "origin"  "seasonl" "compilr" "yrcmpld" "citatin" "source"  "dst_cmm" "version"
# [12] "shp_lng" "shape_r"
  
  ### Individual files call the id number 'iucn_sid'; bli is named sisid
  ### 'iucn_sid' as well;
  ### fix that here!
  names(spp_dbf_info)[names(spp_dbf_info) == 'sisid'] <- 'iucn_sid'

  ### if 'binomial', set to 'sciname'
  if(!'sciname' %in% names(spp_dbf_info)) {
    names(spp_dbf_info)[names(spp_dbf_info) == 'binomil'] <- 'sciname'
  }
  
  ### Some files (e.g. bli and individual files) don't have 'subpop':
  ### add it in as NAs.
  if(!'subpop' %in% names(spp_dbf_info)) spp_dbf_info$subpop <- NA
  
  ## fix presenc colname in bli to be presence
  names(spp_dbf_info)[names(spp_dbf_info) == 'presenc'] <- 'presence'
  
  ### some files don't have 'presence' field: if not present or NA,
  ### set to presence = 1.  Some have presence = 0?
  if(!'presence' %in% names(spp_dbf_info)) spp_dbf_info$presence <- 1
  
  spp_dbf_info <- spp_dbf_info %>%
    mutate(presence = ifelse(presence == 0, 1, presence))


map_info_raw_bli <- spp_dbf_info %>%
  cbind(bli_dbfs) %>%
  dplyr::select(shp_iucn_sid = iucn_sid, sciname, presence, subpop, "dbf_file" = "bli_dbfs")


## now rbind
map_info_raw <- rbind(map_info_raw_iucn, map_info_raw_bli)

### Notes on presence, origin, seasonal fields:
### * presence = 5 is extinct; 4 = probably extinct; others are extant-ish or 
###   uncertain. We will drop field and include all polygons for now.
### * origin is native, introduced, etc.  We will drop this field and not
###   worry about origin.
### * seasonal is breeding/non-breeding/passage.  We will drop this field
###   and not worry about seasonality.

# subspp <- map_info_raw %>%
#   filter(!is.na(subpop)) %>%
#   dplyr::select(shp_iucn_sid, presence) %>%
#   distinct()
### No subspecies vary by IUCN code; drop field for simplicity
### and include all at species level

map_info_sans_dupes <- map_info_raw %>%
  group_by(shp_iucn_sid) %>%
  filter(!(str_detect(dbf_file, 'MARINEFISH|REPTILE|REEF_FORMING_CORALS_PART2|REEF_FORMING_CORALS_PART1|MANGROVES') & 
             length(unique(dbf_file)) > 1)) %>%
  filter(!(shp_iucn_sid == 196026 & str_detect(dbf_file, 'SEASNAKE'))) %>%
    ### in REPTILES, SEASNAKES, and HAGFISH - ditch any dupes
  ungroup() %>%
  distinct()
### NOTE: still has terrestrial mammals and reptiles in it...

subpops_match <- read_csv(file.path(dir_setup, 'int',
                                    sprintf('subpops_match_api_to_shp_%s.csv', 
                                            api_version)),
                          col_types = 'dccdc')

map_info_add_subpops <- map_info_sans_dupes %>%
  full_join(subpops_match, by = c('shp_iucn_sid', 'sciname', 'subpop' = 'shp_subpop')) %>%
  mutate(iucn_sid = ifelse(is.na(api_iucn_sid), shp_iucn_sid, api_iucn_sid)) %>%
  dplyr::select(shp_iucn_sid, iucn_sid, sciname, presence, subpop, dbf_file) %>%
  distinct()

### filter to marine habitat species (to drop terrestrial reptiles e.g.)
marine_spp_ids <- read_csv(file.path(dir_data,
                                     sprintf('spp_marine_from_api_%s.csv', 
                                             api_version)),
                           col_types = 'dcc') %>%
  dplyr::select(iucn_sid, max_depth)

marine_map_info <- map_info_add_subpops %>%
  inner_join(marine_spp_ids, by = 'iucn_sid') %>%
  mutate(comp_assessed = iucn_sid %in% comp_spp$iucn_sid)

comp_assess_check <- marine_map_info %>%
  mutate(dbf_short = basename(dbf_file)) %>%
  group_by(dbf_short, comp_assessed) %>%
  summarize(n_in_gp = n()) %>%
 group_by(dbf_short) %>%
  mutate(both = n() == 2)

### at this point, fix a few "comprehensively assessed" issues:
### * add turtles as comprehensive
### * some species from Spatial Data Download groups do not
###   show up as comprehensively assessed despite text on the IUCN site
turtles <- c('Dermochelys', 'Caretta', 'Lepidochelys', 
             'Natator', 'Chelonia', 'Eretmochelys') %>%
  paste0(collapse = '|')
assessed_gps <- c('BONEFISH', 'CORALS', 'SEACUCUMBERS') %>%
  paste0(collapse = '|')
  ### NOTE: Damselfish removed at recommendation of Gina Ralph
  ### Others, assume that availability of maps on Spatial Data Download Page
  ### is indication of comprehensive assessment despite API call

marine_map_info <- marine_map_info %>%
  mutate(comp_assessed = ifelse(str_detect(dbf_file, assessed_gps), 
                                TRUE, comp_assessed),
         comp_assessed = ifelse(str_detect(sciname, turtles), 
                                TRUE, comp_assessed))

write_csv(marine_map_info, 
          file.path(dir_data,
                    sprintf('spp_marine_maps_%s.csv', api_version)))

```

``` {r generate table of species}

marine_map_info <- read_csv(file.path(dir_data,
  sprintf('spp_marine_maps_%s.csv', api_version))) %>%
  dplyr::select(-presence) %>%
  distinct()

taxa_info <- read_csv(file.path(dir_goal_anx, 'iucn', 
                                'spp_info_from_api_2022-1.csv')) %>%
  dplyr::select(iucn_sid, kingdom:family)

cat_info <- read_csv(file.path(dir_data, 
                               'iucn_risk_current_2022-1.csv')) %>%
  dplyr::select(iucn_sid, cat_score) %>%
  filter(!is.na(cat_score))

included_maps <- marine_map_info %>%
  filter(comp_assessed) %>%
  filter(iucn_sid %in% cat_info$iucn_sid) %>%
  dplyr::select(iucn_sid, shp_iucn_sid, subpop, sciname, dbf_file) %>%
  left_join(taxa_info, by = 'iucn_sid') %>%
  mutate(spp_gp = basename(dbf_file),
         spp_gp = spp_gp %>% tolower() %>% str_replace('.dbf|_part.+', '')) %>%
  mutate(kingdom = tools::toTitleCase(tolower(kingdom)),
         phylum  = tools::toTitleCase(tolower(phylum)),
         order   = tools::toTitleCase(tolower(order)),
         class   = tools::toTitleCase(tolower(class)),
         family  = tools::toTitleCase(tolower(family)))

all_in_the_fam <- included_maps %>%
  group_by(kingdom, phylum, class, order, family, spp_gp) %>%
  summarize(species = length(unique(shp_iucn_sid)),
            subpops = sum(!is.na(subpop)))  

basic_sum <- included_maps %>%
  filter(!str_detect(tolower(class), 'actinopt|mammal|aves|rept')) %>%
  group_by(kingdom, phylum, class, spp_gp) %>%
  summarize(species = length(unique(shp_iucn_sid)),
            subpops = sum(!is.na(subpop)))

nonfish_sum <- included_maps %>%
  filter(str_detect(tolower(class), 'mammal|aves|rept')) %>%
  group_by(kingdom, phylum, class, order, spp_gp) %>%
  summarize(species = length(unique(shp_iucn_sid)),
            subpops = sum(!is.na(subpop)))

fish_sum <- included_maps %>%
  filter(str_detect(tolower(class), 'actinopt')) %>%
  group_by(kingdom, phylum, class, order, family, spp_gp) %>%
  summarize(species = length(unique(shp_iucn_sid)),
            subpops = sum(!is.na(subpop)))

all_sum <- bind_rows(fish_sum, nonfish_sum, basic_sum) %>%
  group_by(spp_gp, kingdom, phylum, class, order) %>%
  summarize(family = paste(unique(family), collapse = ', '),
            species = sum(species),
            subpops = sum(subpops)) %>%
  ungroup() %>%
  mutate(family = ifelse(family == 'NA', NA, family),
         subpops = ifelse(subpops == 0, NA, subpops))

## Check against last year to see if you have all of the same species. 
## 201: We do. 

old_sum <- read_csv(file.path(dir_goal, '_output', 
                             'spp_included_summary.csv'))

setdiff(old_sum$spp_gp, all_sum$spp_gp)

## end check

## write to github
write_csv(all_sum, file.path(dir_goal, '_output', 
                             'spp_included_summary.csv'))
write_csv(all_in_the_fam, file.path(dir_goal, '_output', 
                                    'spp_included_by_family.csv'))
```
